refactoring the way we handle sidebar opening events (#574)
The way `v_summary` handles the loading of the sidebar
assumes there is only one type of sidebar view.

In light of the new "zoomin" view, we would have to extend and
generalize the way we load the sidebar, so as to be able to load both
the authorship and zoomin view using the same mechanism.

The design of the "event bus" is as such. The actual loading of the
tabs is done in the main app, and those functions usually have the form
`updateTabXXX()`, which write the corresponding information object into
`this.tabInfo` and does `this.tabActive='XXX';`. This will then load
the right tab in the sidebar.

To load the sidebar from without another component, the VueJS event
emitting mechanism is used. A function of the form  `openTabXXX()` is
used to call the component's `$emit()`. The main application then
handles the emitted message using the corresponding `updateTabXXX()`.

* Add highlighting to ramps (#544)

To be able to open up a "zoomed-in" view of the ramp, the user must
first be able to select part of the ramp to focus on.

Let's add a way for the user to highlight the range of the ramp to
focus on. This will later translate into the date range for the
"zoomed-in" view to display the relevant commits.

In this particular implementation, we use a global `drags` object as a
way of preventing user from highlighting on multiple ramp charts. E.g.
if a user mousedown on one ramp chart and mouseup on the other, the
second chart will be highlighted with the positions defined between the
two mouse events.

* Merge 'tabs' to get refactored v_authorship

commit b632566535edb678c2b0f685d606e9dacf20e7d0
Author: ongspxm <ongspxm@gmail.com>
Date:   Mon Feb 25 00:46:12 2019 +0800

    added TODO to remove change

commit 4bb21d65c7d75423e94ea6c355f3d6016cf4da94
Author: ongspxm <ongspxm@gmail.com>
Date:   Fri Feb 22 23:38:34 2019 +0800

    updated to use isTabActive instead of tabActive to trigger tab display

commit a7aeefd72eda30f8bac825cadfe2a7c195d90cde
Author: ongspxm <ongspxm@gmail.com>
Date:   Thu Feb 21 22:39:18 2019 +0800

    fix lint

commit b813aa2b98b132a92da552ca4d38895acf89d8a1
Author: ongspxm <ongspxm@gmail.com>
Date:   Thu Feb 21 22:30:35 2019 +0800

    wrap everything within v-authorship

commit 7dbb41be6696d45f532d771c49d3ad45c594d322
Author: ongspxm <ongspxm@gmail.com>
Date:   Thu Feb 21 22:25:45 2019 +0800

    updateCount using document.getElementsByClassName

commit c66a3eeb96cdde3530fc4dc7f3ff8e8958d76a46
Author: ongspxm <ongspxm@gmail.com>
Date:   Thu Feb 21 15:53:22 2019 +0800

    wip. left with updateCount

commit 9e73331ca0ffad2142c811b34806ef6e4583f272
Author: ongspxm <ongspxm@gmail.com>
Date:   Thu Feb 21 15:49:48 2019 +0800

    move expand func to v_authorship

commit 3fcd62b66d0019010395b8dfaf3a0b1735a4d74c
Author: ongspxm <ongspxm@gmail.com>
Date:   Thu Feb 21 15:43:50 2019 +0800

    remove js for button update

commit 4fc11512cb2a2f806034294da3470c396f63b1be
Author: ongspxm <ongspxm@gmail.com>
Date:   Thu Feb 21 15:37:02 2019 +0800

    deactivating the tab

commit a387e19a8b76dfeb4f43d42a64e4b24215f9921f
Author: ongspxm <ongspxm@gmail.com>
Date:   Thu Feb 21 14:51:58 2019 +0800

    wip

* move call emitting to component func

* handle opening of tabs in main app

* refactoring ramps to their own component (#572)

Currently, the entire ramp template resides in `v_summary`. In light of
the "zoomin" tab view, we would have to reuse ramp chart view to
display the commits in the "zoomin" tab.

Let's refactor to move the ramp into its own component, so as to
support reuse in the "zoomin" tab view.

* fix lint

* fix: show min max date on authorship view

* lint

* [#554] Rename 'dashboard' instances to 'report' (#579)

The term 'report' and 'dashboard' are used throughout RepoSense to
refer to the same thing, the result after running analysis.

However, users may not understand the subtle differences between
these two terms. Instead, using a single term would help in
standardization and comprehensibility.

Let's rename all usage of 'dashboard' instances to the term 'report'
instead, as standardized in #220.

* Checkstyle: update addresses of DTD files #586

We are using the DTD files from http://checkstyle.sourceforge.net/
and http://puppycrawl.com/ in our config file of checkstyle.

However, due to security reason, checkstyle decided to remove DTDs from
above websites and ask users to use the DTD files from
https://checkstyle.org/[1].

Let's update the addresses of DTD files correspondingly.

As later version of the suppression DTD file has been released[2],
let's also update the DTD file to the latest version.

[1] Checkstyle removes DTDs from http://checkstyle.sourceforge.net:
checkstyle/checkstyle#6478

[2] Checkstyle Suppressions.xml example:
https://checkstyle.org/config_filters.html#SuppressionFilter_Examples

* [#540] Perform cloning in parallel with analyzing (#560)

Repos are processed sequentially, one at a time.

As cloning does not takes much processing power and analysis does not
take any network bandwidth, the two can be done in parallel to reduce
the total processing time.

Here are some test results to support the above hypothesis:
The test was performed using CS2103 AY1819S1 project repos on my local
machine. The average time taken to generate the report was measured
across 10 runs. Current code took 17 min 21s while new implementation
took 11 min 55s.

Let's clone the next repo in the list while the current repo is being
analyzed.

* [#465] Url: bookmark opened code view (#524)

Our report's url changes along with the state of the dashboard as a
mechanism to allow users to easily revisit her last view as well as
easily be shared with other people.

However, this was only limited to the chart view configuration, nothing
of the code view were part of the state being saved.

To allow users to easily restore their last reviewed code view, let's
also encode any opened code view into the url.

* [#510] Remove unused Checkstyle analysis feature (#597)

The checkstyle analysis feature was added to the code base in the
early stage of RepoSense, when there was some hope of running static
analysis over the code written by authors of the repositories.

As this feature has not been used since early stages(v1.0), and it
is also not in our immediate plans anymore, leaving it will
unnecessarily complicate the codebase.

Let's clean up the code base by removing this unused checkstyle
analysis feature.

* fix: codeview not opening

* update view opeing