[#619] Remove unused classes in functional code (#628)
The MethodAnalyzer and MethodInfo classes were added to the
codebase in the early stage of RepoSense, when there was some
hope of identifying authors of methods of Java file. The
ConfigurationBuilder class was used by GitGrader which has been
remove from codebase and it is not used in the main code.

As these classes have not been used since early stages(v1.0), and they
are also not in our immediate plans anymore, leaving them will
unnecessarily complicate the codebase.

Let's clean up the code base by removing these unused classes.