[#390] repo-config.csv: provide override prefix (#602)
In our implementation, Standalone config has the utmost priority; such
that the user is unable to override any of the fields without disabling
it.

However, there may be cases where user may want to import only certain
fields (e.g. authors) and discard/reconfigure the others. One example
of this can be to import the details of the authors, but change the
file formats of the analysis.

Let's provide the override functionality in the repo-config.csv file
using the prefix "override:".