[#599] CodeView: fix `All` checkbox indication failure (#611)
The `All` checkbox of file formats remains checked even after the user
have performed the deselection of other checkboxes. This is a bug as
the `All` checkbox should have unchecked itself automatically.

This bug originates from #487, where the author have accidentally split
selectFileFormat into two functions. It causes the status of
`All` checkbox gets updated as the separated function was not part of
any execution code.

It is also related to a bug where the `fileFormats` in the frontend
contain formats with 0 lines, which results in the formats displayed
may not be equal to the formats stored in `fileFormats`.

Additionally, the logic of the `All` checkbox is flawed. When the user
selects a checkbox after deselecting all of them, the checkbox will not
update itself correspondingly too.

To restore the intended functionality, let's revert the separation made
in #487 and correct the logic of the `All` checkbox.