[#337] RepoCloner: use single clone to analyze multiple branches of a repository (#625)
When analyzing multiple branches of a repository, RepoSense will clone
that same repository as many times as the number of the branches if
they are not listed consecutively in the config file.

This is inefficient as all the analysis were performed against the same
repo. Hence, there shouldn't be a necessity to clone it multiple times.

Lets use a single clone to analyze multiple branches of a repository
regardless of the order they are listed in the config file.