[#391] GitLsTree: validate filepaths before cloning (#564)
Unix and Windows use a different set of invalid characters, with the
latter being more restrictive.

A valid file path in Unix may be invalid in Windows, which causes git
clone to fail. We classify this error generally as GitCloneException,
which may not be useful in conveying the exact cause of error.

Let's validate file paths and use a more specialized exception to log
invalid file paths, when analysis is done on Windows.