[#601] ChartView: improve the sort controls (#618)
When the ramp charts are grouped together, the sorting is only
performed within the group. This may differs from the expectation of
some users, who may wish to sort them by groups e.g. to list them
accordingly to the group's combined contribution. 

Also, the `reverse` option is not well-presented nor intuitive to the
users enough to understand that it is meant to reverses the sorted
listing. 

Let's fix the mentioned problems by adding another drop-down select
menu to allow the sorting of groups, and include direction arrows in
menu items to replace the need of reverse option.