[#503] CLI: add --version tag to show version of RepoSense (#565)
Users may want to figure out whether the jar file they download is
the correct version of RepoSense they want to use.

The users cannot verify it because we do not have a 'version' tag,
and the generated report do not contain such information.

Let's add a -version tag to show the version of jar file of RepoSense,
and include it in the generated report.

The version can be passed in to RepoSense using -Pversion for both run
and shadowJar. Otherwise, latest git tag will be used instead.