[#603] GitLog: replace author date with commit date (#641)
The ramps in report are placed according to the authored date to
visualize the actual timeline when the authors make their
contribution.

However, GitHub repos use commit date in their timeline. This leads to
discrepancy when performing comparison with the GitHub platform.
Furthermore, this is also the factor causing the url on the ramps
directing users to incorrect dated commit list web page.

To make our product coherent to GitHub users, let's change to use
commit dates for our ramp chart.