[#637] ContributionBar: length changes on breakdown option selected (#638)
Contribution bar length changes when breakdown option is selected.

This happens because multi-line contribution bars are incorrectly
calculated for file formats. If a contribution bar is unable to fit
in the current line, a new line is added and the width of the new line
is subtracted from the previous contribution bar element. However, this
subtraction can lead to a negative number and it is left unchecked.
This will cause the contribution bar to be longer than expected because
the negative width was not accounted for (elements with negative widths
will just be rendered as having zero width).

Lets fix the bar width calculation by checking for negative widths,
and rectify it by removing the element with negative width and
subtracting that amount from the other remaining contribution bar
element.