[#699] Report: standardize the author's name display (#701)
When summary charts are grouped by none or repo/branch, author's name
display shows only the display name. Whereas when grouped by author,
the name display shows the display name as well as GitHub username.

This causes inconsistency in the name display. Furthermore, user won't
be able to refer to the author's GitHub username when the summary
charts are not grouped by author.

Let's standardize all the name display using the format,
[display name] (github username).