[#678] Report: use development version of vue.js (#679)
Vue has a plugin for Chrome and Firefox for debugging Vue.js
applications. However, the minified version of Vue in our frontend is a
production version, so we are unable to use the plugin to debug
RepoSense report.

Let's use the development version of Vue.js to allow the use of Vue
developer tool.