[#689] v_summary.js: remove unused remnants code (#733)
Unused code was found in v_summary.js; specifically code associated
with the variable, filterSortReverse. They are remnants left behind by
the improvement of sort controls introduced in #618.

The unused code was originally bind to the reverse checkbox in the
older version of RepoSense. As the checkbox was replaced by the
introduction of ascending and descending options in #618, the code is
now unused and redundant. This is also partially due to the complexity
brought upon by new sorting group, the reversing has to be done within
the code of descending's logic, rendering those code not reusable.

Let's remove the unused code associated to the variable, 
filterSortReverse and it's implementations.