[#666] CodeView: load un-authored code only when toggled (#740)
Viewing the changes in code of an author causes the app to 
render all authored and un-authored segments within modified files.
A small change in large file will cause the entire file to be preloaded.

When an author has small changes done to a lot of large files, it will
lead to an unresponsive page as loading the large un-authored segments
will affect the rendering speed. Furthermore, the preloading of all 
un-authored code is not necessary as users might only view the 
un-authored code under certain circumstances. (e.g. when the user 
does not understand why the author added the authored code)

Let's modify the app to only load the un-authored code when the 
user toggles the 'Show untouched code' icon.