[#741] Fix view code icon not opening CodeView (#744)
The splitCommitsWeek() method executes a dateRounding() method that 
rounds up the filterSinceDate to the next nearest Monday. When the 
report page is reloaded, the filterSinceDate is set to the value of 
the 'since' hash by the getDates() method. Since within the 
getFiltered() method, splitCommitsWeek() is executed before
getDates(), the filterSinceDate stores an empty string which is 
passed to the dateRounding() method and returns a NaN value. 

Since splitCommitsWeek() method is only called when granularity is 
'week', if users were to set granularity as 'week' and reload the page,
clicking on the view code icon does not open the CodeView panel. 

Let's shift the getDates() method to be executed before 
splitCommitsWeek() to make sure that splitCommitWeek() method
is called with a valid filterSinceDate, which fixes the described bug.