[#711] Analysis: ignore binary files (#728)
Binary files are included in the commits analysis and are shown in the
report view as interpreted plain text.

The produced results are gibberish; junks that do not make sense and
not useful to the readers.

Let's exclude binary files from the analysis.

The method implemented to determine whether a file is binary or not is 
by utilizing "git diff" functionality return a list of committed files 
with the number of lines added and deleted. A file is deemed to be 
binary when the number of lines added and deleted are both "-".