[#680] Vue: fix error message in development mode (#742)
An error message stating "error compiling template: - <template>
cannot be keyed. Place the key on real elements instead." was found in
Chrome browser's console when running Vue in development mode for our
report.

Although the Vue development tool still functions correctly, it 
highlights an issue with the existing style and possibility degrades 
the efficiency of the operations.

Let's remove the Chrome browser console error by shifting the binding 
of keys from the <template> tag to the inner <label> tag.