[#770] Add placeholder for repositories with no author found (#777)
Repositories with no author specified in the configuration files as
well as no commits made during the specified period do not appear in
the report.

This is caused by a TypeError in JavaScript. When an author
configuration can't be found for a repository, RepoSense will perform
the identification of the authors using the commits. However, in the
event that even commits cannot be found, author information will not be
available and result in the error stated above.

As users may not notice the disappearance of these repositories, let's
add a placeholder with an error message of 'NO AUTHOR WITH COMMITS
FOUND WITHIN THIS PERIOD OF TIME' to the affected repositories to
inform users of the occurrence of this error.