[#748] Restore authorship tab (#749)
Reloading the report page with the authorship tab opened will result
to a closed authorship tab.

The info.name property is only created when clicking on the view code
icon and calling openTabAuthorship() method of v_summary. The
initiate() method of v_authorship sets the isTabActive property to be 
false if info.name property does not exist. Reloading the report page 
does not call openTabAuthorship(), which means that the info.name 
property will not exist and the isTabActive is false regardless of 
url value.

Let's modify the initiate() method of v_authorship.js to check for
info.author property instead of info.name as the property exists 
when reloading the page as well as clicking on the view code icon.
This modification properly restores the last viewed authorship tab, 
if any.