[#567] Report#Zip: remove unrelated files and folders (#656)
Zipped files currently contain irrelevant files and/or folders. 

Some users may dump other irrelevant files/folders into 
the same directory as the directory for the repos to be 
cloned into. This results in irrelevant files and folders being 
included in the zip file, which is not intended to be.

Let's zip only the relevant repo folders and the 
corresponding .JSON files into the zip file. 

This method utilizes the existing class `RepoConfiguration` , to
get the list of repos report to be zipped.