[#791] CodeView: bugfix for unexpected colors in line numbers (#792)
The css property of line number only overwrites the color property of 
classes 'hljs-number' and 'hljs-comment'. Others classes such as
'hljs-quote', 'hljs-name' and more are not overwritten.

This will cause the color property of certain line numbers to inherit 
from the hljs classes that are not overwritten.

Let's overwrite all hljs classes to ensure the color property of the 
line number to always be grey.