[#738] CsvParser: fix incorrect splitting of literals commas (#739)
We parse CSV by using comma as a delimiter.

When commas are used in a column as value, by enclosing double quotes,
they are mistakenly treated as a delimiter. This results in incorrect
parsing of values.

Let's use third-party CSV parser library, Apache Commons CSV to
handle the intricacies of CSV parsing in a correct and clean
manner.