[#753] Display individuals' contribution (#762)
Contribution is shown only for each group as a whole.

While it is possible to view the contribution based on each chart by 
hovering over the contribution bar, this requires extra work by the
user. It will be more user-friendly if the contribution can be viewed
on the page itself.

Let's display the contribution for each chart in the chart title.
This information is shown using the same style as the group's total 
contribution.