[#661] Change chart title format (#802)
The chart title uses underscores to separate `org`, `repo` and
`branch`. Furthermore, `displayName` is used for both the chart title
and repo analysis output folder name.

`org`, `repo` and `branch` might have underscores in them, making it
difficult to discern each term. Using a single parameter displayName
to represent both the output folder name and the chart title would
constrain us to not use /. This has shown that the output folder name
and the chart title have little compatibility, and difficulty in coming
up with a coherent format.

Let's use a separate variable, `outputFolderName`, for the out folder
name so that a different format, `org/repo[branch]`, can be used for
the chart title.