[#724] Modify zoom-in feature to based on selected granularity (#729)
The zoom-in feature opens the zoom panel to show the selected
sub-range of a ramp chart in terms of 'commit'. This feature, however, 
only works when the granularity is set to commit or day.

Users will not be able to use this feature with granularity set to week,
as the method getUserCommits() gets all user commits by interating 
through an undefined key object 'dayCommits', which stops the method
execution.

Moreover, long sub-range of ramp charts will contain dense information 
with ramp slices overlapping each other. This happens even when 
granularity is set to day or week, which makes it difficult for users to 
notice substantial info on these ramp charts.

Let's correctly implement the zoom-in feature when granularity is set to
week. Furthermore, let's modify the feature for the ramp chart within 
zoom panel to be in terms of the granularity selected, which allow users 
to observe selected granularity of commits more clearly within the 
zoomed-in range.