[#723] Report: fix modifying dates for Safari users (#801)
Safari browser users are unable to modify sinceDate and untilDate in
the report.

This is due to the datepickers used are HTML5 elements; they are not
supported by Safari. Furthermore, the alternative to modify the date
fields, through the keydown events, are disabled.

Let's allow Safari browser users to use a text-type input for their
sinceDate and untilDate inputs. They can modify the dates by
keying the numeric date values into the date fields followed by hitting
the key, `enter`.