[#778] ChartView: add zoom tab button (#796)
In addition to zooming into a range of ramp slices, the zoom tab also
shows the breakdown of commits that are within the zoom range. This 
functionality however, is not encompassed for the entire ramp chart.

Let's add a button that opens the zoom tab for the entire ramp chart.