[#759] Cloning: remove redundant cloning from remote (#765)
In Windows OS, each repo will be bare cloned from the remote, checked 
for any illegal characters in the file path, then cloned from the 
remote again to be analyzed.

The double cloning of the same repo from the remote increases analysis 
time significantly, especially when the user has a sub-optimal
internet bandwidth.

In addition, when setting repo-config file, if the user specifies a
repo branch that contains file path with illegal characters in the repo
to be analyzed first before the other branches of the same repo, an
exception will be thrown for the first branch with files containing
illegal characters in the repo, and the other branches of the same repo
will be ignored.

Let's: -
- Locally clone from the previously cloned bare repo instead to
reduce the overall time of cloning and analysis.
- Check the other branches of the same repo for files containing 
illegal characters as well as analyze them.

Each repo will be bare cloned. Then for each branch to be analyzed, 
the bare repo will be used to check for any illegal characters in the 
file path. If no illegal characters are found, perform a full clone 
from the bare repo into a new location and directly check out into the 
targeted branch to be analyzed.