[#662] ChartView: add merge group option (#752)
Within ChartView, a group has multiple rows of ramp charts that belongs
to the group itself.

It may be difficult for users to analyse the contribution of the entire
group or to compare the contribution between groups.

Let's add a merge group checkbox such that when the checkbox is
checked, all ramp charts of the same group will be merged together
to form a single ramp chart. This allows users to easily analyse 
the total contribution on the group level.