[#846] CodeView: fix show ramps for this period button (#847)
Clicking on show ramp chart for this period button in the commits
panel calls updateDateRange(), which recomputes the duration of the
ramp chart according to the drags by the user. The method proceeds 
to call deactivateAllOverlays(), which removes the drags by the user.

Since no more drags are left after the first execution of 
updateDateRange(), clicking on the show ramp chart for this period
button will not be able to activate updateDateRange().

Let's use the zoom since date and until date to recompute the duration
of the ramp chart, instead of using the drags. This will ensure that
clicking on the button will always show the ramp chart for the 
designated period.