[#754] Url: include date hashes only when modified (#757)
The initial URL of the report page captures a start and until 
hash parameter even when the users did not specify the start date
and end date fields in the CLI.

The URL is then shared to others. If there are further updates to the 
report, the URL will still navigate to the report page with ramps up 
until the original start date and until date instead of the most
updated dates which may include more recent/older commits.

Let’s modify the report’s initial URL to not capture the start and end 
date hashes when users did not specify the respective date field in
the CLI. Without the start date or end date hash parameter in the URL,
the URL will navigate to the report with the most updated start date
and end date.