[#842] Fix out of bound ramp slices (#844)
Ramp slices appear outside of the ramp charts in week granularity.

This is due to the implementation of splitCommitsWeek() does not take
into account when the duration between since date and until date is
less than one week. When the duration is less than a week, the
algorithm calculates commits up to the following Monday, even if they
are beyond the until date. Hence, it causes a distortion in the ui.

Let's revamp the logic within splitCommitsWeek to check, and trim any
commits beyond the until date.