[#767] ChartView: clear selected zoom region on settings change (#775)
The zoom region on the ramp chart is not cleared when settings in the
toolbar are changed. 

This could be misleading for users as the commits currently in the 
zoom region on the ramp chart do not correspond to those in the
commits panel.

Let's clear the ramp chart's zoom region by calling 
`deactivateAllOverlays()` when the settings are changed.