[#825] ReportGenerator: fix zip paths (#826)
Due to a change in the RepoConfiguration display name in a recent PR,
#802, the ZIP utility now attempts to zip report folders using the new 
display name. Hence, as the name of the report folder is now different 
from the display name, this results in none of the files being zipped
and produce a zip file unusable to display the report.

Let's fix this by directing the zip algorithm to look for the
`OutputFolderName` instead.