Zoom Tab: revamp header ui (#848)
The zoom tab header remains the same irregardless of the grouping
option (authors, repos or none). The period of the zoom tab is also not
shown in the zoom tab.

This creates a problem when the merge group option is used; a merged
group may consists of more than one chart, and the title used in the
header might not apply to all the ramps within the merged chart.

Let's modify the UI of the zoom tab title to use a new sequence:
`group title` -> `chart title` -> `period of zoom tab` ->
`granularity`, and hide the chart title when the merge group option is
checked.

Let's also modify the authorship's title format to match with the zoom
tab's title format as well.