[#703] Log: show analysis progress (#837)
The progress in analyzing the repos is not shown.

Users who are waiting for the analysis to complete may want
to know how many repos RepoSense has analyzed so far.

Let's show the queue number of the repo that is currently
being analyzed. Repos that failed to clone or checkout will
not cause the queue number to increase.