[#808, #815] Docs: update documentation (#822)
With the recently merged pull requests, UI and behavioral changes in 
the report view are not reflected in the documentation.

Users who want to view an author's codes or expand the codes that 
were untouched by an author, may not understand how to do so as the 
user guide relating to these features is outdated.

Let's update the user guide to reflect the changes made to the summary 
and code panel. Let's also take this chance to update components in  
the developer guide that are outdated and can cause confusion with the 
new developers.