[#817] Cypress: add code view loading time benchmark test (#824)
We do not have a benchmark for the time taken to load code view. 

This makes it difficult to ascertain whether new code added will slow
down its opening, which will result in a less positive user experience.

Let's add a benchmark test for the opening of the code view. The
threshold for the load time is 8s, with the buffer time being
the larger of 3s or 10% of the threshold. If out of 5 trials, the
load times of all trials exceed the maximum time, the test is deemed to
have failed.