[#851] Fix url directing to wrong repository when merge group (#854)
When the merge group option is checked, charts are grouped by
authors and granularity is set to 'commit', the url of some commits 
may point to the wrong repository.

This is due to our current structure of the merged groups, where each
merged author stores only one repository name even though they may 
have contributed to more than one repository. This causes all url 
of commits to point to the same repository.

Let's bind the repository name to each commit when merge group 
option is checked to ensure that all url of commits point to the 
correct repository.