Security patch: update cypress to 3.4.1 (#814)
A potential security vulnerability exists in one of our indirect
dependencies, lodash, which requires immediate update to version 4.17.13
or later. As cypress is our only dependency that utilizes the vulnerable
version of lodash, updating it will hence resolve the potential security
issue.

Let's update cypress to version 3.4.1, which is the earliest version
that uses lodash version 4.17.13 or later.