[#520] Group: allow customization of file categories (#787)
"Group" is a new type of classification introduced in this PR to 
classify files/directories that match the given globs into different 
categories. This is similar to how we use the file formats to perform 
breakdown in the ChartView as well as search and filter in CodeView.

This Group functionality provides the ability to customize each 
category, giving users more control over the information reflected in 
the report. To utilize this feature, users simply have to specify the 
categories and the corresponding globs for the classification. 
E.g.: the user can categorize the entire "src/" folder as code, which 
is especially useful when the codebase consists of more than 1 file 
format.

Let's allow users to specify their own custom groups for files in a 
repository, and view a breakdown of contribution by these groups in 
the report.

The user can specify a config file named "groups-config.csv", which 
contains the repository's location, group names and globs to include 
for the group. 

If no group is assigned to a file, it is assigned the default 
group: `other`.

If groups are specified by the user, the Code View and the Contribution
bar will show a breakdown by groups for the checkboxes. Else, breakdown
by file formats are shown instead.

As the code for Format and Grouping are largely similar, let's,
* replace the existing Format class with a new class "FileType",
which uses file globs to determine if a file matches the formats or
groups glob.
* introduce a new class "FileTypeManager" that holds two list of 
FileType: "Format" and "Group".