[#859] Logging: include branch name in repo not found warning (#860)
The message '[WARNING] Repository SOME_REPO is not found in
repo-config.csv.' appears when a corresponding configuration for
`author-config` cannot be found in `repo-config`. 

However, it does not include the branch name, of the SOME_REPO, that
cannot be found in the repo-config.csv. This can confuse the users as
they may not know the exact error since there can be multiple lines of
configuration targeting different branches of one repository.

Let's include the branch name in the warning message to provide users
better clarity of the error.