ChartView: modify the sorting of the authors to use display names (#880)
The sorting of the author by title is according to the authors' 
GitHub usernames.

Github usernames are strings that users have no control over, 
an unconfigurable field. Therefore, sorting according to it does not
have much value to the user, should she chose to configure the display
names.

Let's modify the sorting of author names to be according to the
authors' display names.