[#185] Config, Test: use realistic display names (#875)
The display name of the authors' used for the demo and tests are plain.

They do not cover all the possible allowable use cases provided to
the users. This can cause bugs to go unnoticed in the test environment,
and the actual UI may deviates from our expectation.

Let's rename them to cover more actual scenarios. This includes long
names, names with special characters and numbers, and whitespace.