Report: fix breakdown by file type (#889)
Breakdown by file types functionality is broken when grouping by NONE 
or by AUTHORS.

This is caused by a typo in a variable name that does not exist. As a 
result, the report is unable to populate the list of file types and 
users won't be able to view the breakdown of contribution by file type 
when grouping by NONE or by AUTHORS.

Let's fix the typo in the variable name.