[#886] Appveyor: fix failing unit test (#888)
Appveyor CI is failing due to GitRevParseTest which checks for
existence of master branch.

This is because the git rev-parse command in the failing test method
has been using the wrong directory; the reference should be using the
target repo's root. The test, however, uses the repo folder name (which
is parent of the target repo's root) and cause the test to fail.

Lets fix the test to use the correct directory.

When not on the target repo's root, git tends to use the resources of
higher order repositories (if it is found). On AppVeyor, RepoSense has
always been cloned as a repository, which subsequently clones the
target repositories into one of its subfolders for analysis. Hence, the
issue has been kept hidden as the verification has been performed
against the RepoSense's resources and not the target's resources.

This issue came to the light in the release commit when AppVeyor clones
the RepoSense as a release, which does not contain the master branch,
causing the git rev-parse to fail.