[#893] Report#ContributionBar: fix incorrect width (#896)
When breakdown by file format is selected, the width of the
contribution bars are computed incorrectly for some file types.

This occurs because the current implementation for calculating the
contribution bars does not keep track of the bar width of the current
line properly. In addition, the current implementation has rather
convoluted logic.

Lets fix and improve the logic for calculating contribution bar width
for file types.