[#918] Bugfix: add catch-all clause on analysis of each repo (#919)
In #917, an exception during the analysis one of the CS2103 student
repositories caused the report generation to fail for other
repositories as well.

This occurred because the bug led to an exception that was not being
caught.

Even after #917 is fixed, there might still be other undiscovered
bugs which can lead to uncaught exceptions. Lets add a catch-all
clause during the analysis of a single repository so that uncaught
exceptions will not affect other repositories.