[#806, #951] ReportGenerator: return paths of generated files (#939)
The method `ReportGenerator#generateReposReport` returns a list of
paths to the JSON files that need to be zipped. 

However, the list is made up by the file paths of expected output and
not the file paths of actual generated files. As a good practice, and
also to ensure that only the correct files are zipped, the latter is
preferred. 

Let's return a Optional<Path> from `FileUtil#writeJsonFile` so that
callers of this method can determine whether a JSON file was
successfully generated.

This allows `generateReposReport` to only return paths to JSON files
that were successfully generated.