[#871, #952] CSV files: ignore blank lines before the header (#903)
CsvParser treats the first non-empty line as the header row.

When a CSV file contains blank line(s) above the header, the blank line
would be taken as the header, causing the subsequent rows in the CSV
file to be flagged as malformed. Hence, any CSV file with blank lines
before the header will be parsed incorrectly. 

Let's modify CsvParser such that it skips all the blank lines preceding
the header, and picks the first non blank line as the header.

The Apache Commons CSV library currently does not provide any methods
to skip blank lines preceding the header. This limitation is documented
in:
https://issues.apache.org/jira/browse/CSV-127 

Hence, let's implement a custom line-skipping behavior using a
BufferedReader.