[#934] CsvParser: Remove extra comma in warning logs (#935)
When a line in a CSV config is found to be malformed, the logger prints 
the contents of the line in a warning message. The current method causes
an extra comma to be printed.

This extra comma may make debugging more confusing.

Let's modify the method used to extract the contents of a line, such 
that there are no extra commas.