[#865] AuthorConfig: further improve the branch field (#878)
In the author-config.csv, branch is an optional field that will
automatically uses the default branch in the repo field.

However, this can be improved such that if the branch field is not
filled, we can bind that AuthorConfig to all the RepoConfig that has the
same repo's location irregardless of branch. This can help users to
easily setup config files to analyse multiple branches of a repo.

Let's bind the AuthorConfig to all the RepoConfig that has the same
repo's location irregardless of branch if the branch field is not
filled.